feat: use tekton Pipeline and add JSFunctionBuild (#20)
* feat: rebuild the runtime when a function changes
* feat: add version number tags to the runtime image
* feat: use tekton Pipeline and add JSFunctionBuild

Fixes: https://github.com/openshift-cloud-functions/js-function-operator/issues/13

This change adds a new CRD - `JSFunctionBuild` whose responsibility
is to run a Tekton `Pipeline` when it is provided with a new revision
of the `JSFunction` container image URL. The controller manages a
Tekton `PipelineResource` for each function. When the function changes
and a new image/tag is needed, the controller updates the `PipelineResource`
with the new image tag, and then initiates a build.